constbackend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
python-dateutil==2.8.2 